Extracting unrelated log functionality out of LicenseMatchLogger class.
