+++ src/lib/engine/license-matching/score-logger.ts
@@ -1,14 +1,10 @@
 import { CsvStream } from '../../data/csv';
-import DataDir from '../../data/dir';
 import { License } from '../../marketplace/model/license';
 import { shorterLicenseInfo } from './license-grouper';
 
 export class LicenseMatchLogger {
 
-  #scoreStream;
-  constructor(private logDir: DataDir, stream: CsvStream) {
-    this.#scoreStream = stream;
-  }
+  constructor(private stream: CsvStream) { }
 
   l1!: License;
   l2!: License;
@@ -27,9 +23,9 @@ export class LicenseMatchLogger {
     const score = this.score;
     if (score > 0) {
       const reason = this.scores.filter(([s, r]) => s).map(([s, r]) => `${r}=${s}`).join(',');
-      this.#scoreStream.writeObjectRow({ score, reason, ...shorterLicenseInfo(this.l1) });
-      this.#scoreStream.writeObjectRow({ score, reason, ...shorterLicenseInfo(this.l2) });
-      this.#scoreStream.writeBlankRow();
+      this.stream.writeObjectRow({ score, reason, ...shorterLicenseInfo(this.l1) });
+      this.stream.writeObjectRow({ score, reason, ...shorterLicenseInfo(this.l2) });
+      this.stream.writeBlankRow();
     }
   }
 
@@ -38,29 +34,4 @@ export class LicenseMatchLogger {
     this.scores.push([score, reason]);
   }
 
-  logMatchResults(matches: License[][]) {
-    const groups = matches.map(group => group.map(shorterLicenseInfo));
-
-    this.logDir.file('matched-groups-all.csv').writeCsvStream(allMatchGroupsLog => {
-      this.logDir.file('matched-groups-to-check.csv').writeCsvStream(checkMatchGroupsLog => {
-        for (const match of groups) {
-          for (const shortLicense of match) {
-            allMatchGroupsLog.writeObjectRow(shortLicense);
-          }
-          allMatchGroupsLog.writeBlankRow();
-
-          if (match.length > 1 && (
-            !match.every(item => item.tech_email === match[0].tech_email) ||
-            !match.every(item => item.company === match[0].company) ||
-            !match.every(item => item.tech_address === match[0].tech_address)
-          )) {
-            for (const shortLicense of match) {
-              checkMatchGroupsLog.writeObjectRow(shortLicense);
-            }
-            checkMatchGroupsLog.writeBlankRow();
-          }
-        }
-      });
-    });
-  }
 }
